Addrecipe moved to the right place
