+++ database/db.json
@@ -1,6 +1,6 @@
 {
   "user": {
-    "name": "Filip"
+    "name": "filip"
   },
   "recipes": [
     {
