+++ src/js/components/layout/MainView/Dashboard/AddRecipe/index.js
@@ -1,50 +1,87 @@
 import React from 'react';
-
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faPlusSquare } from "@fortawesome/free-solid-svg-icons";
 
 
 const AddRecipe = () => {
-
-  const onSumbit=(e)=>{
+  const onSubmit = e => {
     e.preventDefault();
   };
-
 
   return (
     <div className="add__recipe__container">
       <div className="add__recipe__header">
         <h1>new recipe</h1>
-        <button classname="add__recipe__button">
-          save and close
-        </button>
+        <button classname="add__recipe__button">Save and close</button>
       </div>
 
       <div className="add__recipe__input">
-        <form onSubmit={()=>onSumbit}>
+        <form onSubmit={() => onSubmit}>
           <label>
-            name
-          <input type="text" value="Submit"/>
+            Recipe's name
+            <input
+              className="recipe__name-input"
+              type="text"
+              placeholder="add recipe's name"
+            />
           </label>
           <label>
-            description
-          <input type="text" value="Submit"/>
-        </label>
+            Recipe's description
+            <textarea
+              rows="6"
+              className="recipe__decsription-input"
+              type="text"
+              placeholder="add recipe's description"
+            />
+          </label>
         </form>
       </div>
 
-      <div className="add__recipe__content">
-        <div className="add__recipe__content-intructions">
-          <ul className="instructions-list">
-            <li className="instructions-element">Turn on the oven</li>
-          </ul>
+      <div className="add__recipe__content-container">
+        <div className="add__recipe__content">
+          <div className="add__recipe__content-header">
+            <h1>instructions</h1>
+            <div className="add__recipe__content-fill">
+              <form onSubmit={() => onSubmit}>
+                <label>
+                  <textarea
+                    rows="5"
+                    placeholder="add a new instruction"
+                    type="text"
+                  />
+                </label>
+              </form>
+              <FontAwesomeIcon className="icon-1" icon={faPlusSquare} />
+            </div>
+          </div>
+          <div className="add__recipe__content-list">
+            <ul className="add__recipe__content-list">
+              <li className="instructions-element">Turn on the oven</li>
+            </ul>
+          </div>
         </div>
-        <div className="add__recipe__content-ingridients">
-          <ul className="ingridients-list">
-            <li className="ingridients-element">Couliflower 2</li>
-          </ul>
+
+        <div className="add__recipe__content">
+          <div className="add__recipe__content-header">
+            <h1>ingredients</h1>
+            <div className="add__recipe__content-fill">
+              <form onSubmit={() => onSubmit}>
+                <label>
+                  <textarea placeholder="add an new ingredient" type="text" />
+                </label>
+              </form>
+              <FontAwesomeIcon className="icon-2" icon={faPlusSquare} />
+            </div>
+          </div>
+          <div className="add__recipe__content-list">
+            <ul>
+              <li className="ingridients-element">Couliflower 2</li>
+            </ul>
+          </div>
         </div>
       </div>
     </div>
   );
 };
 
-export default AddRecipe; 
+export default AddRecipe;
